Change lib name to sot-dynamic and dynamic plugin name to dynamic
