+++ src/CMakeLists.txt
@@ -30,7 +30,7 @@ SET(integrator-force-rk4_plugins_dependencies integrator-force)
 SET(integrator-force-exact_plugins_dependencies integrator-force)
 
 
-FOREACH(lib ${libs})
+FOREACH(lib ${plugins})
   ADD_LIBRARY(${lib} SHARED ${lib}.cpp)
 
   SET_TARGET_PROPERTIES(${lib} PROPERTIES
@@ -61,15 +61,15 @@ FOREACH(lib ${libs})
 ENDFOREACH(lib)
 
 # Main Library
-ADD_LIBRARY(dynamic SHARED dynamic.cpp)
-TARGET_LINK_LIBRARIES(dynamic jrl-dynamics)
-TARGET_LINK_LIBRARIES(dynamic sot-core)
-TARGET_LINK_LIBRARIES(dynamic dynamic-graph)
-PKG_CONFIG_USE_DEPENDENCY(dynamic jrl-dynamics)
-PKG_CONFIG_USE_DEPENDENCY(dynamic sot-core)
-PKG_CONFIG_USE_DEPENDENCY(dynamic dynamic-graph)
+ADD_LIBRARY(${LIBRARY_NAME} SHARED sot-dynamic.cpp)
+TARGET_LINK_LIBRARIES(${LIBRARY_NAME} jrl-dynamics)
+TARGET_LINK_LIBRARIES(${LIBRARY_NAME} sot-core)
+TARGET_LINK_LIBRARIES(${LIBRARY_NAME} dynamic-graph)
+PKG_CONFIG_USE_DEPENDENCY(${LIBRARY_NAME} jrl-dynamics)
+PKG_CONFIG_USE_DEPENDENCY(${LIBRARY_NAME} sot-core)
+PKG_CONFIG_USE_DEPENDENCY(${LIBRARY_NAME} dynamic-graph)
 
-INSTALL(TARGETS dynamic DESTINATION ${CMAKE_INSTALL_LIBDIR})
+INSTALL(TARGETS ${LIBRARY_NAME} DESTINATION ${CMAKE_INSTALL_LIBDIR})
 
 
 # Install empty __init__.py files in intermediate directories.
